refactor: reorganize tests directory for better structure and maintainability
