+++ tests/unit/layers/test_attention.py
@@ -1,6 +1,7 @@
 import math
 
 import pytest
+pytestmark = pytest.mark.unit
 import torch
 
 from energy_transformer.layers.attention import MultiHeadEnergyAttention
