+++ tests/unit/layers/test_hopfield.py
@@ -1,4 +1,5 @@
 import pytest
+pytestmark = pytest.mark.unit
 import torch
 
 from energy_transformer.layers.hopfield import HopfieldNetwork
