+++ tests/unit/models/test_base.py
@@ -1,4 +1,5 @@
 import pytest
+pytestmark = pytest.mark.unit
 import torch
 
 from energy_transformer.layers.base import (
