+++ tests/unit/models/vision/test_viet.py
@@ -1,4 +1,5 @@
 import pytest
+pytestmark = pytest.mark.unit
 import torch
 from torch import nn
