+++ tests/unit/spec/test_library.py
@@ -7,6 +7,7 @@ including validation, context handling, and composition.
 from dataclasses import fields
 
 import pytest
+pytestmark = pytest.mark.unit
 
 from energy_transformer.spec.library import (
     ClassificationHeadSpec,
